refactor : put vec helpers in separate file
