+++ include/exec/helper-gen-common.h
@@ -15,6 +15,10 @@
 #include "exec/helper-gen.h.inc"
 #undef  HELPER_H
 
+#define HELPER_H "accel/tcg/tcg-runtime-sym-vec.h"
+#include "exec/helper-gen.h.inc"
+#undef  HELPER_H
+
 #define HELPER_H "accel/tcg/plugin-helpers.h"
 #include "exec/helper-gen.h.inc"
 #undef  HELPER_H
